% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_img_data.R
\name{download_img_data}
\alias{download_img_data}
\title{Download T1 and T2 data}
\usage{
download_img_data(lib.loc = NULL)
}
\arguments{
\item{lib.loc}{a character vector with path names of R libraries. 
Passed to \code{\link{img_data}}}
}
\value{
Logical indicator if the files were downloaded.
}
\description{
Download T1 and T2 data for Examples
}
