% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{s.hist}
\alias{s.hist}
\title{smoothed_histogram}
\format{A GAM from mgcv for x and y from histograms}
\usage{
s.hist
}
\description{
Smoothed histogram of image
}
\examples{
\dontrun{
data(t2.voi)
img.hist = hist(t2.voi,
breaks=2000,
plot=FALSE)
y = img.hist$counts
x = img.hist$mids
x = x[!is.na(y)];
y = y[!is.na(y)]
# 70 used for speed of example
s.hist = smooth_hist(x, y, k=70)
}
}
\keyword{datasets}

