% Generated by roxygen2 (4.0.1): do not edit by hand
\name{smooth_hist}
\alias{smooth_hist}
\title{Histogram smoothing for whitestripe}
\usage{
smooth_hist(x, y, deg = 4, k = floor(min(250, length(x)/2)),
  method = "REML", ...)
}
\arguments{
\item{x}{values of midpoints from \code{\link{hist}}}

\item{y}{values of counts from \code{\link{hist}}}

\item{deg}{degree of polynomials used}

\item{k}{Number of knots}

\item{method}{Method for smoothing for GAM}

\item{...}{Arguments passed to \code{\link{gam}}}
}
\value{
List of objects: x and y coordinates of histogram, coefficients from GAM,
fitted values from GAM, the GAM model, the knots fittted, and degrees of polynomials
}
\description{
Uses a generalized additive model (GAM) to smooth a
histogram for whitestripe
}
\examples{
\dontrun{
data(t2.voi)
img.hist = hist(t2.voi,
breaks=2000,
plot=FALSE)
y = img.hist$counts
x = img.hist$mids
x = x[!is.na(y)];
y = y[!is.na(y)]
# 70 used for speed of example
s.hist = smooth_hist(x, y, k=70)
}
}
\seealso{
gam
}

