% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{t1.voi.hist}
\alias{t1.voi.hist}
\title{t1.voi.hist}
\format{A volume of interest histogram from a T1 image for smoothing}
\usage{
t1.voi.hist
}
\description{
Histogram of VOI of T1 template image
}
\examples{
\dontrun{
if (download_img_data()){
t1 = readNIfTI(system.file("T1Strip.nii.gz", package="WhiteStripe"))
t1.voi = make_img_voi(t1)
any(is.na(t1.voi))
# FALSE
t1.voi.hist = hist(t1.voi,
breaks=2000,
plot=FALSE)
#save(t1.voi.hist, file="data/t1.voi.hist.rda", compress = TRUE,
# compression_level=9)
}
}
}
\keyword{datasets}

