% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{t1.voi}
\alias{t1.voi}
\title{t1.voi}
\format{A volume of interest from a T1 image for smoothing}
\usage{
t1.voi
}
\description{
VOI of T1 template image
}
\examples{
\dontrun{
t1 = readNIfTI(system.file("T1Strip.nii.gz", package="WhiteStripe"))
t1.voi = make_img_voi(t1)
any(is.na(t1.voi))
# FALSE
#save(t1.voi, file="data/t1.voi.rda", compress = TRUE,
# compression_level=9)
}
}
\keyword{datasets}

