% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1.voi.R
\docType{data}
\name{t2.voi.hist}
\alias{t2.voi.hist}
\title{Histogram of VOI of T2 template image}
\format{
A histogram volume of interest from a T2 image for smoothing
}
\usage{
t2.voi.hist
}
\description{
Histogram of VOI of T2 template image
}
\examples{
\dontrun{
lib.loc = tempdir() 
if (download_img_data(lib.loc = lib.loc)){
t2 = readNIfTI(system.file("T2Strip.nii.gz", package="WhiteStripe",
lib.loc = lib.loc))
t2.voi = make_img_voi(t2)
any(is.na(t2.voi))
# FALSE 
t2.voi.hist = hist(t2.voi, 
breaks=2000, 
plot=FALSE)  
#save(t2.voi.hist, file="data/t2.voi.hist.rda", compress = TRUE,
# compression_level=9) 
}
} 
}
\keyword{datasets}
