% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_img_data.R
\name{ws_img_data}
\alias{ws_img_data}
\title{Return Filenames of T1 and T2 data}
\usage{
ws_img_data(lib.loc = NULL, warn = TRUE)
}
\arguments{
\item{lib.loc}{a character vector with path names of R libraries. 
Passed to \code{\link{system.file}}}

\item{warn}{Should a warning be printed if the images were not
there}
}
\value{
Vector of filenames
}
\description{
Return filenames T1 and T2 data for example 
and vignettes
}
