% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_hist.R
\name{get.deriv.smooth.hist}
\alias{get.deriv.smooth.hist}
\title{Gets nth derivative of smoothed histogram}
\usage{
get.deriv.smooth.hist(x, coefs, knots, deg = 4, deriv.deg = 1)
}
\arguments{
\item{x}{values from smooth_hist}

\item{coefs}{Coefficients from GAM from smooth_hist}

\item{knots}{Number of knots fit for GAM}

\item{deg}{Degree of polynomials}

\item{deriv.deg}{The degree of the derivative.}
}
\value{
Derivative of smoothed histogram
}
\description{
This function outputs the nth derivative of a histogram smooth.
}
\examples{
data(smoothed_histogram)
dy<-get.deriv.smooth.hist(xvals, 
coefs=s.hist$coefs,
knots=s.hist$knots,
deg=s.hist$deg,
deriv.deg=1)
}
