\name{get.largest.mode}
\alias{get.largest.mode}
\title{Grab largest peak}
\usage{
get.largest.mode(x, y, ...)
}
\arguments{
  \item{x}{values of midpoints from \code{\link{hist}}}

  \item{y}{values of counts from \code{\link{hist}}}

  \item{...}{arguments to be passed to
  \code{\link{smoooth.hist}}}
}
\value{
Value of x that is the largest peak
}
\description{
This function grabs the largest peak of the histogram
}
\examples{
\dontrun{

}
}

