% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whitestripe_ind.R
\name{make_img_voi}
\alias{make_img_voi}
\title{Make Image VOI}
\usage{
make_img_voi(img, slices = 80:120, na.rm = TRUE, ...)
}
\arguments{
\item{img}{Image (T1 usually or T2).  Array or object of class nifti}

\item{slices}{Slices to take for the image voi}

\item{na.rm}{Remove NAs from mean.  This is for double checking}

\item{...}{Arguments passed from other methods (not used)}
}
\value{
VOI of image.
}
\description{
Creates a VOI of Image for the specified slices
}

