% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{s.hist}
\alias{s.hist}
\title{smoothed_histogram}
\format{A GAM from mgcv for x and y from histograms}
\usage{
s.hist
}
\description{
Smoothed histogram of image
}
\examples{
\dontrun{
data(t2.voi.hist)
y = t2.voi.hist$counts
x = t2.voi.hist$mids
x = x[!is.na(y)];
y = y[!is.na(y)]
# 70 used for speed of example
s.hist = smooth_hist(x, y, k=70)
}
}
\keyword{datasets}

