% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1.voi.R
\docType{data}
\name{s.hist}
\alias{s.hist}
\title{Smoothed histogram of image}
\format{
A GAM from \code{mgcv} for x and y from histograms
}
\usage{
s.hist
}
\description{
Smoothed histogram of image
}
\examples{
\dontrun{ 
data(t2.voi.hist)
y = t2.voi.hist$counts
x = t2.voi.hist$mids
x = x[!is.na(y)];
y = y[!is.na(y)]
# 70 used for speed of example
s.hist = smooth_hist(x, y, k=70)
}
}
\keyword{datasets}
