% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{t2.voi}
\alias{t2.voi}
\title{t2.voi}
\format{A volume of interest from a T2 image for smoothing}
\usage{
t2.voi
}
\description{
VOI of T2 template image
}
\examples{
\dontrun{
if (download_img_data()){
t2 = readNIfTI(system.file("T2Strip.nii.gz", package="WhiteStripe"))
t2.voi = make_img_voi(t2)
any(is.na(t2.voi))
# FALSE
#save(t2.voi, file="data/t2.voi.rda", compress = TRUE,
# compression_level=9)
}
}
}
\keyword{datasets}

