% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitestripe_ind.R
\name{whitestripe_norm}
\alias{whitestripe_norm}
\title{Normalize Image using white stripe}
\usage{
whitestripe_norm(img, indices, ...)
}
\arguments{
\item{img}{Array or object of class nifti}

\item{indices}{Indices of white stripe from \code{\link{whitestripe}}.
Can also be a mask (indices where mask > 0 are used.)}

\item{...}{arguments to be passed to \code{\link{mean}} and \code{\link{sd}}}
}
\value{
Object of same class as \code{img}, but normalized
}
\description{
Taking the indices from white stripe to normalize the
intensity values of the brain
}
